[#549] Sample: Refactor data module packages
