+++ sample-compose/app/src/main/java/co/nimblehq/sample/compose/di/modules/RepositoryModule.kt
@@ -1,7 +1,7 @@
 package co.nimblehq.sample.compose.di.modules
 
+import co.nimblehq.sample.compose.data.remote.services.ApiService
 import co.nimblehq.sample.compose.data.repositories.RepositoryImpl
-import co.nimblehq.sample.compose.data.services.ApiService
 import co.nimblehq.sample.compose.domain.repositories.Repository
 import dagger.Module
 import dagger.Provides
