+++ sample-compose/data/src/main/java/co/nimblehq/sample/compose/data/local/storages/BaseSharedPreferences.kt
@@ -1,4 +1,4 @@
-package co.nimblehq.sample.compose.data.storages
+package co.nimblehq.sample.compose.data.local.storages
 
 import android.content.SharedPreferences
