+++ sample-compose/data/src/main/java/co/nimblehq/sample/compose/data/remote/models/responses/ErrorResponse.kt
@@ -1,4 +1,4 @@
-package co.nimblehq.sample.compose.data.responses
+package co.nimblehq.sample.compose.data.remote.models.responses
 
 import co.nimblehq.sample.compose.domain.models.Error
 import com.squareup.moshi.Json
