+++ sample-compose/data/src/main/java/co/nimblehq/sample/compose/data/repositories/RepositoryImpl.kt
@@ -1,8 +1,8 @@
 package co.nimblehq.sample.compose.data.repositories
 
 import co.nimblehq.sample.compose.data.extensions.flowTransform
-import co.nimblehq.sample.compose.data.responses.toModels
-import co.nimblehq.sample.compose.data.services.ApiService
+import co.nimblehq.sample.compose.data.remote.models.responses.toModels
+import co.nimblehq.sample.compose.data.remote.services.ApiService
 import co.nimblehq.sample.compose.domain.models.Model
 import co.nimblehq.sample.compose.domain.repositories.Repository
 import kotlinx.coroutines.flow.Flow
