--- sample-compose/data/src/main/java/co/nimblehq/sample/compose/data/services/ApiService.kt
@@ -1,10 +1,0 @@
-package co.nimblehq.sample.compose.data.services
-
-import co.nimblehq.sample.compose.data.responses.Response
-import retrofit2.http.GET
-
-interface ApiService {
-
-    @GET("users")
-    suspend fun getResponses(): List<Response>
-}
