+++ sample-compose/data/src/test/java/co/nimblehq/sample/compose/data/test/MockUtil.kt
@@ -1,6 +1,6 @@
 package co.nimblehq.sample.compose.data.test
 
-import co.nimblehq.sample.compose.data.responses.ErrorResponse
+import co.nimblehq.sample.compose.data.remote.models.responses.ErrorResponse
 import io.mockk.every
 import io.mockk.mockk
 import okhttp3.ResponseBody
@@ -31,7 +31,7 @@ object MockUtil {
     )
 
     val responses = listOf(
-        co.nimblehq.sample.compose.data.responses.Response(
+        co.nimblehq.sample.compose.data.remote.models.responses.Response(
             id = 1,
             userName = "name1",
         )
